refactor(Benefit): 총혜택 금액과 혜택 내역을 구하는 기능을 리팩토링
변경 내용:
- 경로를 변경함(Model 폴더 내로 이동)
- 테스트 코드에 경로 변경을 반영
